Refactor :app ui package slightly
