+++ app/src/main/java/com/ricknout/rugbyranker/ui/matches/WorldRugbyMatchDateItemDecoration.kt
@@ -1,4 +1,4 @@
-package com.ricknout.rugbyranker.ui.common
+package com.ricknout.rugbyranker.ui.matches
 
 import android.content.Context
 import android.graphics.Canvas
