+++ app/src/main/java/com/ricknout/rugbyranker/ui/matches/WorldRugbyMatchPagedListAdapter.kt
@@ -1,4 +1,4 @@
-package com.ricknout.rugbyranker.ui.common
+package com.ricknout.rugbyranker.ui.matches
 
 import android.view.LayoutInflater
 import android.view.View
