+++ app/src/main/java/com/ricknout/rugbyranker/ui/rankings/RankingsFragment.kt
@@ -13,7 +13,6 @@ import com.ricknout.rugbyranker.R
 import dagger.android.support.DaggerFragment
 import javax.inject.Inject
 import androidx.core.view.isVisible
-import com.ricknout.rugbyranker.ui.common.WorldRugbyRankingListAdapter
 import androidx.core.os.bundleOf
 import androidx.recyclerview.widget.DividerItemDecoration
 import androidx.work.WorkInfo.State
